pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "serialized_data_interface"
version = "0.3.0"
description = "Serialized Data Interface for Juju Operators"
authors = [
    "Dominik Fleischmann <dominik.fleischmann@canonical.com>",
    "Kenneth Koski <kenneth.koski@canonical.com>"
]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/canonical/serialized-data-interface/"
repository = "https://github.com/canonical/serialized-data-interface/"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License"
]

[tool.poetry.dependencies]
python = "^3.8"

jsonschema = "3.2"
ops = "^1.1"
pyyaml = "5.3"
requests = "2.25"

[tool.poetry.dev-dependencies]
pytest = "6"
flake8 = "3.8"
black = "20.8b1"
